PlotSpectrumBase in lib-builtin-effects
